Compile I2C functions with C linkage
The BMI270-Sensor-API and BMM150-Sensor-API libraries are implemented in C, and therefore expect all function pointers be compiled with C linkage. The Arduino compiler coincidentally uses the same calling convention for static member functions and `extern "C"` functions as of now, but this could easily change in the future.